Use the new implementation of getAllCombinationsOfMinAndMaxLengthWithCallback that allocates less memory
The usage of getAllCombinations in canPlaySpellOnCurve generates all the combinations of lands and then filter them according to the card CMC which can be inefficient when the CMC is lower than the number of lands to test. We can optimize this by using getAllCombinationsOfMinAndMaxLengthWithCallback2 instead that will generate the require number of combination.